+++ pystac/extensions/label.py
@@ -4,13 +4,13 @@ https://github.com/stac-extensions/label
 """
 
 from enum import Enum
-from pystac.extensions.base import ExtensionManagementMixin
+from pystac.extensions.base import ExtensionManagementMixin, SummariesExtension
 from typing import Any, Dict, Iterable, List, Optional, Set, Union, cast
 
 import pystac
 from pystac.serialization.identify import STACJSONDescription, STACVersionID
 from pystac.extensions.hooks import ExtensionHooks
-from pystac.utils import get_required
+from pystac.utils import get_required, map_opt
 
 SCHEMA_URI = "https://stac-extensions.github.io/label/v1.0.0/schema.json"
 
@@ -50,6 +50,28 @@ class LabelType(str, Enum):
 
     ALL = [VECTOR, RASTER]
     """Convenience attribute for checking if values are valid label types"""
+
+
+class LabelTask(str, Enum):
+    """Enumerates recommended values for "label:tasks" field."""
+
+    def __str__(self) -> str:
+        return str(self.value)
+
+    REGRESSION = "regression"
+    CLASSIFICATION = "classification"
+    DETECTION = "detection"
+    SEGMENTATION = "segmentation"
+
+
+class LabelMethod(str, Enum):
+    """Enumerates recommended values for "label:methods" field."""
+
+    def __str__(self) -> str:
+        return str(self.value)
+
+    AUTOMATED = "automated"
+    MANUAL = "manual"
 
 
 class LabelClasses:
@@ -121,6 +143,15 @@ class LabelClasses:
             ",".join([str(x) for x in self.classes])
         )
 
+    def __eq__(self, o: object) -> bool:
+        if isinstance(o, LabelClasses):
+            o = o.to_dict()
+
+        if not isinstance(o, dict):
+            return NotImplemented
+
+        return self.to_dict() == o
+
     def to_dict(self) -> Dict[str, Any]:
         """Returns the dictionary representing the JSON of this instance."""
         return self.properties
@@ -180,6 +211,15 @@ class LabelCount:
         """Returns the dictionary representing the JSON of this instance."""
         return self.properties
 
+    def __eq__(self, o: object) -> bool:
+        if isinstance(o, LabelCount):
+            o = o.to_dict()
+
+        if not isinstance(o, dict):
+            return NotImplemented
+
+        return self.to_dict() == o
+
 
 class LabelStatistics:
     """Contains statistics for regression/continuous numeric value data.
@@ -233,6 +273,15 @@ class LabelStatistics:
     def to_dict(self) -> Dict[str, Any]:
         """Returns the dictionary representing the JSON of this LabelStatistics."""
         return self.properties
+
+    def __eq__(self, o: object) -> bool:
+        if isinstance(o, LabelStatistics):
+            o = o.to_dict()
+
+        if not isinstance(o, dict):
+            return NotImplemented
+
+        return self.to_dict() == o
 
 
 class LabelOverview:
@@ -379,6 +428,15 @@ class LabelOverview:
         """Returns the dictionary representing the JSON of this LabelOverview."""
         return self.properties
 
+    def __eq__(self, o: object) -> bool:
+        if isinstance(o, LabelOverview):
+            o = o.to_dict()
+
+        if not isinstance(o, dict):
+            return NotImplemented
+
+        return self.to_dict() == o
+
 
 class LabelExtension(ExtensionManagementMixin[pystac.Item]):
     """A class that can be used to extend the properties of an
@@ -403,8 +461,8 @@ class LabelExtension(ExtensionManagementMixin[pystac.Item]):
         label_type: LabelType,
         label_properties: Optional[List[str]] = None,
         label_classes: Optional[List[LabelClasses]] = None,
-        label_tasks: Optional[List[str]] = None,
-        label_methods: Optional[List[str]] = None,
+        label_tasks: Optional[List[Union[LabelTask, str]]] = None,
+        label_methods: Optional[List[Union[LabelMethod, str]]] = None,
         label_overviews: Optional[List[LabelOverview]] = None,
     ) -> None:
         """Applies label extension properties to the extended Item.
@@ -499,28 +557,28 @@ class LabelExtension(ExtensionManagementMixin[pystac.Item]):
             self.obj.properties[CLASSES_PROP] = classes
 
     @property
-    def label_tasks(self) -> Optional[List[str]]:
+    def label_tasks(self) -> Optional[List[Union[LabelTask, str]]]:
         """Gets or set a list of tasks these labels apply to. Usually a subset of 'regression',
         'classification', 'detection', or 'segmentation', but may be arbitrary
         values."""
         return self.obj.properties.get(TASKS_PROP)
 
     @label_tasks.setter
-    def label_tasks(self, v: Optional[List[str]]) -> None:
+    def label_tasks(self, v: Optional[List[Union[LabelTask, str]]]) -> None:
         if v is None:
             self.obj.properties.pop(TASKS_PROP, None)
         else:
             self.obj.properties[TASKS_PROP] = v
 
     @property
-    def label_methods(self) -> Optional[List[str]]:
+    def label_methods(self) -> Optional[List[Union[LabelMethod, str]]]:
         """Gets or set a list of methods used for labeling.
 
         Usually a subset of 'automated' or 'manual', but may be arbitrary values."""
         return self.obj.properties.get("label:methods")
 
     @label_methods.setter
-    def label_methods(self, v: Optional[List[str]]) -> None:
+    def label_methods(self, v: Optional[List[Union[LabelMethod, str]]]) -> None:
         if v is None:
             self.obj.properties.pop("label:methods", None)
         else:
@@ -654,6 +712,83 @@ class LabelExtension(ExtensionManagementMixin[pystac.Item]):
             raise pystac.ExtensionTypeError(
                 f"Label extension does not apply to type {type(obj)}"
             )
+
+    @staticmethod
+    def summaries(obj: pystac.Collection) -> "SummariesLabelExtension":
+        """Returns the extended summaries object for the given collection."""
+        return SummariesLabelExtension(obj)
+
+
+class SummariesLabelExtension(SummariesExtension):
+    """A concrete implementation of :class:`~SummariesExtension` that extends
+    the ``summaries`` field of a :class:`~pystac.Collection` to include properties
+    defined in the :stac-ext:`Label Extension <label>`.
+    """
+
+    @property
+    def label_properties(self) -> Optional[List[str]]:
+        """Get or sets the summary of :attr:`LabelExtension.label_properties` values
+        for this Collection.
+        """
+
+        return self.summaries.get_list(PROPERTIES_PROP)
+
+    @label_properties.setter
+    def label_properties(self, v: Optional[List[LabelClasses]]) -> None:
+        self._set_summary(PROPERTIES_PROP, v)
+
+    @property
+    def label_classes(self) -> Optional[List[LabelClasses]]:
+        """Get or sets the summary of :attr:`LabelExtension.label_classes` values
+        for this Collection.
+        """
+
+        return map_opt(
+            lambda classes: [LabelClasses(c) for c in classes],
+            self.summaries.get_list(CLASSES_PROP),
+        )
+
+    @label_classes.setter
+    def label_classes(self, v: Optional[List[LabelClasses]]) -> None:
+        self._set_summary(
+            CLASSES_PROP, map_opt(lambda classes: [c.to_dict() for c in classes], v)
+        )
+
+    @property
+    def label_type(self) -> Optional[List[LabelType]]:
+        """Get or sets the summary of :attr:`LabelExtension.label_type` values
+        for this Collection.
+        """
+
+        return self.summaries.get_list(TYPE_PROP)
+
+    @label_type.setter
+    def label_type(self, v: Optional[List[LabelType]]) -> None:
+        self._set_summary(TYPE_PROP, v)
+
+    @property
+    def label_tasks(self) -> Optional[List[Union[LabelTask, str]]]:
+        """Get or sets the summary of :attr:`LabelExtension.label_tasks` values
+        for this Collection.
+        """
+
+        return self.summaries.get_list(TASKS_PROP)
+
+    @label_tasks.setter
+    def label_tasks(self, v: Optional[List[Union[LabelTask, str]]]) -> None:
+        self._set_summary(TASKS_PROP, v)
+
+    @property
+    def label_methods(self) -> Optional[List[Union[LabelMethod, str]]]:
+        """Get or sets the summary of :attr:`LabelExtension.label_methods` values
+        for this Collection.
+        """
+
+        return self.summaries.get_list(METHODS_PROP)
+
+    @label_methods.setter
+    def label_methods(self, v: Optional[List[Union[LabelMethod, str]]]) -> None:
+        self._set_summary(METHODS_PROP, v)
 
 
 class LabelExtensionHooks(ExtensionHooks):